refactor: split CLI module by concern
